adding thneed PC support
